Conflicts:
	urapp/URApp/ARJoint/View Controllers/ArMode/ViewController+SetUp.swift
	urapp/URApp/ARJoint/View Controllers/ArMode/ViewController.swift
	urapp/URApp/ARJoint/View Controllers/ChatBot/ChatViewController.swift

+++ urapp/URApp/ARJoint/View Controllers/ChatBot/Movement.swift
@@ -28,6 +28,10 @@ class Movement {
     public static let VENTOSA_OFF = "14"
     public static let CONTINUA = "15"
     public static let PAUSA = "16"
+    public static let ESPERA = "17"
+    public static let DATA = "18"
+    public static let SMALL_TALK = "19"
+    public static let LOGIN = "20"
     
     public static let DIRECTION_UP = "arriba"
     public static let DIRECTION_DOWN = "abajo"
@@ -40,6 +44,30 @@ class Movement {
     public static let AMMOUNT_MUCH = "mucho"
     public static let AMMOUNT_LITTLE = "poco"
     public static let AMMOUNT_DEFAULT = "default"
+    
+    public static let JOINT_BASE = "base"
+    public static let JOINT_WRIST = "muñeca"
+    public static let JOINT_ELBOW = "codo"
+    public static let JOINT_SHOULDER = "hombro"
+    
+    public static let DATA_TEMP = "temperatura"
+    public static let DATA_VOLT = "voltaje"
+    public static let DATA_CORR = "corriente"
+    
+    public static let SMTLK_TIEMPO = "tiempo"
+    public static let SMTLK_TRAFICO = "trafico"
+    public static let SMTLK_NOTICIAS = "noticias"
+    public static let SMTLK_CANCION = "cancion"
+    public static let SMTLK_STORY = "historia"
+    public static let SMTLK_RADIO = "radio"
+    
+    var smalltalk = ["tiempo": "weather",
+                     "trafico": "traffic",
+                     "radio": "radio",
+                     "cancion": "singasong",
+                     "historia":  "tellstory",
+                     "noticias":  "flashbriefing"]
+    
     
     private var com: RobotComunication
     
@@ -71,19 +99,8 @@ class Movement {
     }
     
     func dance() {
-        com.send("def M():\n")
-        com.send("  move = True\n")
-        com.send("  while move:\n")
-        var i = 0;
-        for pose in dance_poses {
-            com.movej_to(Position(pose, "1.8", "1.4", "0", "0", "0"))
-            i += 1;
-            com.send("  while is_steady() == False:\n")
-            com.send("      sleep(0.01)\n")
-            com.send("  end\n")
-        }
-        com.send("  end\n")
-        com.send("end\n")
+        loadProgram("mambo2.urp")
+        
     }
     
     func stopMovement() {
@@ -222,42 +239,44 @@ class Movement {
     func stopProgramming() {
         var order = 1
         
-        let fr = NSFetchRequest<Mov>(entityName: "Mov")
-        let _movements = try! context.fetch(fr)
-        
-        let movements = Array<Mov>()
-        
-        for movement in _movements {
-            if (movement.name == programName) {
-                context.delete(movement)
-            }
-            saveContext()
-
-        }
-        
-        for inst in programInstructions! {
-            let i = NSEntityDescription.insertNewObject(forEntityName: "Mov", into: context) as! Mov
-            i.name = programName
-            i.order = Int16(order)
-            i.ventosa = inst.0
-            if (inst.1 == 1) {
-                //position
-                i.positions = inst.2
-            } else {
-                //wait
-                i.time = inst.2
+        if (programInstructions != nil) {
+            let fr = NSFetchRequest<Mov>(entityName: "Mov")
+            let _movements = try! context.fetch(fr)
+            
+            let movements = Array<Mov>()
+            
+            for movement in _movements {
+                if (movement.name == programName) {
+                    context.delete(movement)
+                }
+                saveContext()
+                
             }
             
-            saveContext()
-            order += 1
+            for inst in programInstructions! {
+                let i = NSEntityDescription.insertNewObject(forEntityName: "Mov", into: context) as! Mov
+                i.name = programName
+                i.order = Int16(order)
+                i.ventosa = inst.0
+                if (inst.1 == 1) {
+                    //position
+                    i.positions = inst.2
+                } else {
+                    //wait
+                    i.time = inst.2
+                }
+                
+                saveContext()
+                order += 1
+            }
+            
+            print(movements)
+            
+            programming = false
+            self.programName = nil
+            self.ventosaStatus = false
+            self.programInstructions = nil
         }
-        
-        print(movements)
-        
-        programming = false
-        self.programName = nil
-        self.ventosaStatus = false
-        self.programInstructions = nil
         setVentosa(false)
     }
     
@@ -285,20 +304,29 @@ class Movement {
     }
     
     func saveInstructionWait (time: Int) {
+        if (programInstructions == nil) {
+            self.programInstructions = Array<(Bool, Int, String)>()
+        }
         programInstructions!.append((ventosaStatus, 2, String(time)))
     }
     
     func saveInstructionPosition () {
         var pos1 = "", pos2 = "."
         
-        while (pos1 != pos2 || pos1 == "" || pos2 == "") {
+        
+        while (pos1 != pos2 || pos1 == "" || pos1 == "Error command: 'actual_qactual_qactual_q' not available") {
             pos1 = pos2
             com.sendData("actual_q")
             pos2 = com.recvData()
+            if (pos1 == "Error command: 'actual_qactual_qactual_q' not available"){
+                print("FAIL")
+            }
             print("Recieving this position: \(pos2)")
-            usleep(2000)
+            usleep(20000)
         }
-        
+        if (programInstructions == nil) {
+            self.programInstructions = Array<(Bool, Int, String)>()
+        }
         programInstructions!.append((ventosaStatus, 1, pos2))
     }
     
@@ -330,5 +358,32 @@ class Movement {
     
     func continueProgram() {
         com.sendCommand("play\n")
+    }
+    
+    func getJson(_ data: String) -> [NSNumber] {
+        com.sendData(data)
+        let received = com.recvRawData()
+        do {
+            let array = try (JSONSerialization.jsonObject(with: received as Data) as? [NSNumber])!
+            return array
+        } catch  {
+            print("Error json \(error)")
+        }
+        return [NSNumber]()
+        
+    }
+    
+    func smalltalk(_ type: String, _ radio: String) {
+        let action = smalltalk[type]
+        print("{\"action\":\"\(action ?? "")\",\"value\":\"\(radio)\"}")
+        com.sendAlexa("{\"action\":\"\(action ?? "")\",\"value\":\"\(radio)\"}")
+    }
+    
+    func stopAlexa() {
+        com.sendAlexa("{\"action\":\"stop\",\"value\":\"\"}")
+    }
+    
+    func sendAlexa(_ message: String) {
+        com.sendAlexa(message)
     }
 }
